Python script for EwE data extraction
This script can be used to extract specified data from Ecopath with Ecosim (EwE) files of the format ".ewemdb". For that, a connection to the EwE core has to be established first, which here is done by importing custom functions from the EwEApi_3 script created by Tushith Islam (https://gitlab.com/tushith) from the Center for Policy Exploration, Analysis and Simulation. Since the latter is not yet publicly available, it has not been uploaded in this repository and if in need, please contact him directly for further questions.